backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
langchain>=0.1.12
langgraph>=0.0.26
langsmith>=0.1.23
langchain-groq>=0.0.1
groq>=0.4.2
pydantic>=2.6.3
pandas>=2.2.1
python-multipart>=0.0.9
python-dotenv>=1.0.1
uvloop>=0.19.0; sys_platform != 'win32'
httpx>=0.27.0
